Rename Duke to PulsePilot and remove data file
