--- src/main/java/storage/DataFile.java
@@ -1,16 +1,0 @@
-package storage;
-
-import health.Health;
-
-public class DataFile {
-    private Health health;
-
-    public void save(int id, Health health) {
-    }
-
-    public static DataFile load(int id) {
-        // return new Storage(...)
-        return null;
-    }
-}
-
