+++ src/test/java/seedu/pulsepilot/PulsePilotTest.java
@@ -1,10 +1,10 @@
-package seedu.duke;
+package seedu.pulsepilot;
 
 import static org.junit.jupiter.api.Assertions.assertTrue;
 
 import org.junit.jupiter.api.Test;
 
-class DukeTest {
+class PulsePilotTest {
     @Test
     public void sampleTest() {
         assertTrue(true);
